# frontend/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "PTSite API",
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api"
        }
    ],
    "security": [
        {
            "http": []
        }
    ],
    "paths": {
...
        "/v1/players/quick-add": {
            "post": {
                "operationId": "player.quickAdd",
                "summary": "Add a first-timer by nickname only, e.g. from the results form. Results keepers and admins",
                "tags": [
                    "Player"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/QuickAddPlayerRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "`PlayerResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/PlayerResource"
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            }
        },
...
        "/v1/players/{player}/thumbnail": {
            "get": {
                "operationId": "player.thumbnail",
                "summary": "A player's thumbnail, as an image file. 404 when the player has none.\nThe answer can be cached for a year: put the player's thumbnail_version in the URL (?v=...)",
                "tags": [
                    "Player"
                ],
                "parameters": [
                    {
                        "name": "player",
                        "in": "path",
                        "required": true,
                        "description": "The player ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "headers": {
                            "Cache-Control": {
                                "schema": {
                                    "type": "string",
                                    "const": "private, max-age=31536000, immutable"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
        "/v1/players/{player}/photo": {
            "get": {
                "operationId": "player.photo",
                "summary": "A player's photo, larger than the thumbnail, as an image file. 404 when the player has none.\nThe answer can be cached for a year: put the player's photo_version in the URL (?v=...)",
                "tags": [
                    "Player"
                ],
                "parameters": [
                    {
                        "name": "player",
                        "in": "path",
                        "required": true,
                        "description": "The player ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "headers": {
                            "Cache-Control": {
                                "schema": {
                                    "type": "string",
                                    "const": "private, max-age=31536000, immutable"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            },
            "post": {
                "operationId": "player.storePhoto",
                "summary": "Set a player's photo and thumbnail from one picture, replacing the ones there were. Admins, and the player\nthemself. The server cuts the picture from the middle to 3 wide by 4 tall and makes both images as JPEG:\nthe photo with 600 x 800 pixels and the thumbnail with 180 x 240.\nSend the picture as multipart/form-data in `image`: JPEG, PNG or WebP, 8 MB at most, and between 180 and\n4096 pixels a side",
                "tags": [
                    "Player"
                ],
                "parameters": [
                    {
                        "name": "player",
                        "in": "path",
                        "required": true,
                        "description": "The player ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/SavePlayerPhotoRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "`PlayerResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/PlayerResource"
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            },
            "delete": {
                "operationId": "player.destroyPhoto",
                "summary": "Remove a player's photo and thumbnail together. Admins, and the player themself",
                "tags": [
                    "Player"
                ],
                "parameters": [
                    {
                        "name": "player",
                        "in": "path",
                        "required": true,
                        "description": "The player ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "`PlayerResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/PlayerResource"
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "http": {
                "type": "http",
                "scheme": "bearer"
            }
        },
        "schemas": {
...
            "PlayerResource": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "nickname": {
                        "type": "string"
                    },
                    "name": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "status": {
                        "type": "string"
                    },
                    "archived": {
                        "type": "boolean"
                    },
                    "thumbnail_version": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "description": "Null when the player has no thumbnail. The image is at /players/{id}/thumbnail; add ?v=<this value>\nto the URL, because the answer is cached for a year."
                    },
                    "photo_version": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "description": "The same for the larger photo, at /players/{id}/photo."
                    },
                    "email": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "description": "Contact details only for admins and the player themself."
                    },
                    "birth_date": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "login": {
                        "type": [
                            "object",
                            "null"
                        ],
                        "description": "Site access, only for admins: null when the player has no login.",
                        "properties": {
                            "username": {
                                "type": "string"
                            },
                            "role": {
                                "type": "string",
                                "enum": [
                                    "player",
                                    "results_keeper",
                                    "admin"
                                ]
                            },
                            "last_login_at": {
                                "type": [
                                    "string",
                                    "null"
                                ]
                            }
                        },
                        "required": [
                            "username",
                            "role",
                            "last_login_at"
                        ]
                    },
                    "updated_at": {
                        "type": [
                            "string",
                            "null"
                        ]
                    }
                },
                "required": [
                    "id",
                    "nickname",
                    "name",
                    "status",
                    "archived",
                    "thumbnail_version",
                    "photo_version",
                    "updated_at"
                ],
                "title": "PlayerResource"
            },
...
            "QuickAddPlayerRequest": {
                "type": "object",
                "properties": {
                    "nickname": {
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 60
                    }
                },
                "required": [
                    "nickname"
                ],
                "title": "QuickAddPlayerRequest"
            },
...
            "SavePlayerPhotoRequest": {
                "type": "object",
                "description": "A picture for a player's photo, sent as multipart/form-data in the field `image`. The limits fit a picture\nstraight from a phone's camera, and keep the image software within the server's memory.",
                "properties": {
                    "image": {
                        "type": "string",
                        "format": "binary",
                        "contentMediaType": "application/octet-stream",
                        "description": "Maximum file size: 8192 kilobytes."
                    }
                },
                "required": [
                    "image"
                ],
                "title": "SavePlayerPhotoRequest"
            },
...
        },
        "responses": {
            "ValidationException": {
                "description": "Validation error",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Errors overview."
                                },
                                "errors": {
                                    "type": "object",
                                    "description": "A detailed description of each field that failed validation.",
                                    "additionalProperties": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                }
                            },
                            "required": [
                                "message",
                                "errors"
                            ]
                        }
                    }
                }
            },
            "AuthenticationException": {
                "description": "Unauthenticated",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            },
            "ModelNotFoundException": {
                "description": "Not found",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            },
...
        }
    }
}